From start to master MAJ 17/11/2021

+++ app/src/main/java/com/openclassrooms/realestatemanager/activity/main/maps/MapsEstateViewState.kt
@@ -1,3 +1,3 @@
 package com.openclassrooms.realestatemanager.activity.main.maps
 
-data class MapsEstateViewState (val id: Int, val lat : Double, val lng: Double)
+data class MapsEstateViewState (val id: Long, val lat : Double, val lng: Double)
